Fix vitals route helper export
