+++ src/app/api/internal/vitals/__tests__/route.test.ts
@@ -1,4 +1,4 @@
-import { sanitizeMetricPayload } from "../route";
+import { sanitizeMetricPayload } from "../sanitizeMetricPayload";
 
 describe("sanitizeMetricPayload", () => {
         beforeEach(() => {
